+++ LegendaryExplorer/LegendaryExplorerCore/UnrealScript/Language/Tree/IntegerLiteral.cs
@@ -6,11 +6,13 @@ namespace LegendaryExplorerCore.UnrealScript.Language.Tree
 {
     public class IntegerLiteral : Expression
     {
-        public int Value;
+        //needs to be a long because this is assigned before it is combined with a preceding minus sign
+        //and the minimum value of an int is of one greater magnitude than the max value
+        public long Value;
 
         public string NumType = INT;
 
-        public IntegerLiteral(int val, int start = -1, int end = -1) 
+        public IntegerLiteral(long val, int start = -1, int end = -1) 
             : base(ASTNodeType.IntegerLiteral, start, end)
         {
             Value = val;
